requirements.txt
qrcode==7.4.2
pillow==10.0.1
python-barcode==0.15.1
flask==2.3.3
flask-cors==4.0.0
brother-ql==0.9.4
pyusb==1.2.1
python-dotenv==1.0.0